Update package reference for Spectre.Console and enhance song flushing confirmation. and author on the render works

+++ Jammer.Core/src/Songs.cs
@@ -9,6 +9,13 @@ namespace Jammer
         {
             if (Directory.Exists(Preferences.songsPath))
             {
+                string v = Message.Input("Are you sure you want to Recursively delete '" + Preferences.songsPath + "'? (y/n)", "Flush Jammer Songs", true);
+                if (v != "y")
+                {
+                    AnsiConsole.Clear();
+                    AnsiConsole.MarkupLine($"[red]Jammer songs flush cancelled.[/]");
+                    return;
+                }
                 Directory.Delete(Preferences.songsPath, true);
                 AnsiConsole.MarkupLine($"[green]Jammer songs flushed.[/]");
